JDBC-3988 DBN OJVM Support (side observations - unjustified long attribute holder for data-type length)

+++ src/main/java/com/dbn/database/common/metadata/def/DBDataTypeMetadata.java
@@ -34,7 +34,7 @@ public interface DBDataTypeMetadata {
     @ObjectIdentifier
     String getDeclaredTypeProgram() throws SQLException;
 
-    long getDataLength() throws SQLException;
+    int getDataLength() throws SQLException;
 
     int getDataPrecision() throws SQLException;
