+++ src/main/java/com/dbn/database/common/metadata/impl/DBDataTypeMetadataImpl.java
@@ -53,8 +53,8 @@ public class DBDataTypeMetadataImpl extends DBObjectMetadataBase implements DBDa
         return getString(prefix + "DECL_TYPE_PROGRAM");
     }
 
-    public long getDataLength() throws SQLException {
-        return resultSet.getLong(prefix + "DATA_LENGTH");
+    public int getDataLength() throws SQLException {
+        return resultSet.getInt(prefix + "DATA_LENGTH");
     }
 
     public int getDataPrecision() throws SQLException {
